revert 4 commits
Solution of performing regex by chunks left matching chunks undecideable.
It was a good proposition at first, because it can make stack overflow
impossible for any string, removing extra checks making it faster and
also not having to take strlen(), and not having to go though entire
string if the match is at the beginning. However, if we split string
at 256 char for example, the character at index 256 are at the boundary
where we stop matching, thus we loose the state and it cannot match the word
or any pattern at that 256 offset. If not this, it would of been a perfect
solution and maybe the fastest regex implementation out there?