build.sh: cleanup
Forking the compile steps to the background then waiting 1 second for
each file not compiled was removed, because with this project this
wasted time.

Support for DESTDIR and PREFIX was added.

CC now defaults to cc.

Each function no longer exits, which allows functions to be joined
together: so if the script is invoked with $1 as install and the
executable hasn't been built yet, install() will call build().

ERREXIT is now also set.