sbuf: massive refactor
This commit changes the string api
down to a macro and enable use of
pointers freely from anywhere. The
point is to not deny or hinder
development. Restrictions shall not
exist, it should be clear and straight
forward to see how the "thing" is to
be implemented.
new sbufn_ functions terminate strings.

Few other changes:
	- lbuf_rd now tries to read file up to
	1MB, if it's bigger it shall adjust itself
	to fit arbitary sized file. Reading in bigger
	chunks is faster.
	- lbuf_linecount shall be computed once on the
	file read.
	- Removed static buffer for autocomplete suggestions
	it's size can get arbitrary large.

The patch is extensive but most things tested by now.
you'll notice exe size increase by ~15K due to sbuf
inlining, and compiler taking advantage of optimization.
Though primary goal was not optimization, think of it
like icing on the cake.