sort out all the '\n' termination mess+++
The commit brings in the long awaited refactor of what
'\n' character actually does.
- The flag REG_NEWLINE is now in use, and changes the
regex termination character from '\0' to '\n'. There is
a clear distinction and RATIONALE when one is used
over the other.
- It is now clear and well defined the meaning of $
character in syntax hl
- Any ex_prints or related strings to the status bar
no longer need '\n'. This was a complete dumpsterfire
as the prompt did not use '\n', but some messages did,
it did not affect anything, except create incoherent
noise. It was very convoluted how regexp with $ been
at work there, where if $ removed it could get stuck
in infinite loops.
- Syntax hilight terminates at '\0'. This makes it
possible to do hl past '\n' which in some situations
is needed.
- Fix well known bug when cursor can mismatch text if
text is in reverse and goes past the screen + being
set at the leftmost screen. It was caused by poor handling
of '\n'. There is still an issue where it will not hilight
pattern at eol, because the culled string gets padded and
reversed again changing the physical location of '\0'.
Not a big deal, possible to fix, probably not worth.
- ressurect vi_wait() message, it's not so much about
the message though, it is there to let you know when
command finishes executing. It was realized after removal.
- ex command "ft" now shall reload the hl ft, may be
used to clear out left overs from dynamic highlights
(after being disabled)
- :se nohl will not just set ft to '/' and call it a
day, now it will disable all syntax hilight for real.