led: fix valid, yet invalid logical write to lbuf
See, this is why we refactored sbuf. So bugs
like this can get fixed in 0 lines of code.

It was making an invalid logical write into
underlying lbuf memory replacing the '\n'
Deletion of '\n' is invalid for some operations
where '\n' assumed to exist. Most of the time
it probably was harmless, but it did break
some possible operations.