Gave const qualifier to some functions.
Not only is that makes it easier for the compiler
to optimize the concerned functions but this
could help track some bugs. The main motivation
to do so was that the function `ex_command` had
to manipulate the output of the getenv function.
This output should not be modified (source:
`man 3 getenv`).