28. fix ^v limitations
Even though the project is said to be complete, i still make
changes if I find some genius way to improve it from the
technical perspective. This time fix ^v for my asian friends.
Who knows, in the future, based on how I solved this led_render();
function will be changed as well to use some better approach
on culling the strings. (the most difficult problem in nextvi)