vi: remove memory cleanup
Generally, the OS will cleanup the memory asynchronously upon
application exit which is faster than doing it in userland.
If the number of allocations is high, which is a possibility,
the cleanup code can hold the application for couple of seconds
which is not good for the user (really annoying).
This patch makes it instant in any case.
Cleaning up memory on exit is useful only for debugging, however
when development is completed can be removed. Therefore a special
patch will be created on patches branch, in case somebody wants
to debug leaks.