ex: posixify, e! reedits and discards changes
This is not 100% compliant, since just e will still
be used to open unnamed buffer(s).
It's a tradeoff, but luckily for us there is no
difference between e and e! except the saved check,
so it's an okay decision to make; keeping both
features in tact, but with a quirk.