vi: scrap pbuf for full savepath (for fssearch)
Well, it worked but it's not full idea since pbuf
cannot switch to the same buffer when it gets
reloaded. I also want to have a global mark
work independently from fssearch so I subdivided
the number-space for arguments to create 2 separate
marks. We won't need more marks so this fits perfectly.