lbuf: fix neatvi's marks
Neatvi does a poor job implementing marks. When lbuf is undone
and redone, and the marks are created in region,
the positions were lost when it could still be valid and
preserved.

[ ] marks did not play well with undo either, implement
them in lopt system instead. Now redo and undo will not
break them.

The patch should also improve general editing performance
by ironing out by proccessing marks in a single chunk.