vi.c: fix insert extends xoff calculation
xoff wasn't calculated correctly if pref grows or shrinks
using ^d and ^t from insert. It's harmless, but would break
backspace functionality. Fix it. Actually, vi doesn't need
to handle the character deletions on the line anymore as
insert could do it after this change, yet I did not go
with it as it would make backspace undo less granular.