led.c: improve recursive insert behaviors
Few things as follows:
1. Correct ft after v/ and ^b switching.
2. Prevent niche crash potential via :bx or :e relocation in
recursive state.
3. Bind ^o to the buffer so that recursive operation can return
to original insert position and ^l can redraw correctly in addition
to correct ft set. The idea is to be able to postpone vi/ex insert
mode at any point and come back to it later even if the buffer
was switched. This is possible now because Q or :@Q exists and
can be used instead when state tracking not wanted.
4. Unlocked more efficient memory reuse scenario when exiting
insert.